Update package.json
To install your plugin with ionic/cordova we need to execute ionic cordova plugin add danielsogl-cordova-plugin-clipboard, the problem is that the plugin's name saved in our package.json is the id you set in your package.json > cordova > id (com.danielsogl.cordova.clipboard). So When we fork the project and restore plugins, it try to do ionic cordova plugin add com.danielsogl.cordova.clipboard and it fails... The other problem is if we want to remove the plugin because ionic cordova plugin rm danielsogl-cordova-plugin-clipboard fails, we have to do ionic cordova plugin add com.danielsogl.cordova.clipboard.

By changing the cordova's id by 'danielsogl-cordova-plugin-clipboard', this will improve the installation/unistallation of it and facilitate its implementation for all those who use it.